# compaction: recalculate pacing on each beat