--- a/src/lsm/compaction.zig
+++ b/src/lsm/compaction.zig
@@ -315,10 +315,10 @@ pub fn CompactionType(
             /// Number of beats we should aim to finish this compaction in. It might be fewer, but
             /// it'll never be more.
             beats_max: ?u64,
+            beats_finished: u64 = 0,
             /// The total number of source values for this compaction.
             /// This is fixed for the duration of the compaction.
             compaction_tables_value_count: u64,
-            value_count_per_beat: u64 = 0,
 
             // The total number of source values processed by this compaction across the bar. Must
             // equal compaction_tables_value_count by bar_apply_to_manifest(). Tracked
@@ -396,6 +396,7 @@ pub fn CompactionType(
             };
 
             grid_reservation: Grid.Reservation,
+            value_count_per_beat: u64,
 
             // TODO: This is now always 0 / 1 so get rid of it and just use index_read_done?
             index_blocks_read_b: usize = 0,
@@ -725,14 +726,9 @@ pub fn CompactionType(
             const bar = &compaction.bar.?;
             assert(!bar.move_table);
 
-            assert(bar.value_count_per_beat == 0);
+            assert(bar.beats_max == null);
 
-            // TODO: Move this calculation into beat_grid_reserve, and subtract the values we've
-            // already done from it.
-            // This way we self correct our pacing and better spread the work out....
-            bar.value_count_per_beat = stdx.div_ceil(bar.compaction_tables_value_count, beats_max);
-            assert(bar.value_count_per_beat > 0);
-
+            bar.beats_max = beats_max;
             bar.target_index_blocks = target_index_blocks;
             assert(target_index_blocks.count > 0);
 
@@ -740,11 +736,9 @@ pub fn CompactionType(
             // be null!
             bar.source_a_immutable_block = source_a_immutable_block;
 
-            log.debug("bar_setup_budget({s}): bar.compaction_tables_value_count={} " ++
-                "bar.value_count_per_beat={}", .{
+            log.debug("bar_setup_budget({s}): bar.compaction_tables_value_count={}", .{
                 compaction.tree_config.name,
                 bar.compaction_tables_value_count,
-                bar.value_count_per_beat,
             });
         }
 
@@ -766,13 +760,26 @@ pub fn CompactionType(
             // If we're move_table, only the manifest is being updated, *not* the grid.
             assert(!bar.move_table);
 
-            assert(bar.value_count_per_beat > 0);
+            assert(bar.beats_max != null);
+
+            // Calculate how many values we have to compact each beat, to self-correct our pacing.
+            // Pacing will have imperfections due to rounding up to fill target value blocks and
+            // immutable table filtering duplicate values.
+            const beats_remaining = bar.beats_max.? - bar.beats_finished;
+            const value_count_per_beat = stdx.div_ceil(
+                bar.compaction_tables_value_count - bar.source_values_merge_count,
+                beats_remaining,
+            );
+            assert(bar.compaction_tables_value_count > bar.source_values_merge_count);
+            assert(beats_remaining > 0);
+            assert(bar.source_values_merge_count + value_count_per_beat * beats_remaining >=
+                bar.compaction_tables_value_count);
 
             // The +1 is for imperfections in pacing our immutable table, which might cause us
             // to overshoot by a single block (limited to 1 due to how the immutable table values
             // are consumed.)
             const value_blocks_per_beat = stdx.div_ceil(
-                bar.value_count_per_beat,
+                value_count_per_beat,
                 Table.layout.block_value_count_max,
             ) + 1;
 
@@ -788,15 +795,18 @@ pub fn CompactionType(
             // (actually, we want to still panic but with something nicer like vsr.fail)
             const grid_reservation = compaction.grid.reserve(total_blocks_per_beat).?;
             log.debug("beat_grid_reserve({s}): total_blocks_per_beat={} " ++
-                "index_blocks_per_beat={} value_blocks_per_beat={}", .{
+                "index_blocks_per_beat={} value_blocks_per_beat={} " ++
+                "beat.value_count_per_beat={} ", .{
                 compaction.tree_config.name,
                 total_blocks_per_beat,
                 index_blocks_per_beat,
                 value_blocks_per_beat,
+                value_count_per_beat,
             });
 
             compaction.beat = .{
                 .grid_reservation = grid_reservation,
+                .value_count_per_beat = value_count_per_beat,
             };
         }
 
@@ -1268,14 +1278,14 @@ pub fn CompactionType(
                 // a full value block.
                 source_exhausted_bar = bar.source_values_merge_count ==
                     bar.compaction_tables_value_count;
-                source_exhausted_beat = beat.source_values_processed >= bar.value_count_per_beat;
+                source_exhausted_beat = beat.source_values_processed >= beat.value_count_per_beat;
 
                 log.debug("blip_merge({s}): beat.source_values_processed={} " ++
-                    "bar.value_count_per_beat={}. (source_exhausted_bar={}, " ++
+                    "beat.value_count_per_beat={}. (source_exhausted_bar={}, " ++
                     "source_exhausted_beat={})", .{
                     compaction.tree_config.name,
                     beat.source_values_processed,
-                    bar.value_count_per_beat,
+                    beat.value_count_per_beat,
                     source_exhausted_bar,
                     source_exhausted_beat,
                 });
@@ -1855,6 +1865,7 @@ pub fn CompactionType(
             compaction.grid.forfeit(beat.grid_reservation);
 
             // Our beat is done!
+            bar.beats_finished += 1;
             compaction.beat = null;
         }
 
@@ -1909,6 +1920,11 @@ pub fn CompactionType(
             // Assert we've written all the values we've merged.
             // TODO: Can we assert target_values_merge_count > 0 here?
             assert(bar.target_values_merge_count == bar.target_values_write_count);
+
+            // Assert we've finished within the number of beats we were allocated.
+            // TODO(metric): Track the delta between target and actual.
+            if (!bar.move_table)
+                assert(bar.beats_finished <= bar.beats_max.?);
 
             // Mark the immutable table as flushed, if we were compacting into level 0.
             if (compaction.level_b == 0 and bar.table_info_a.immutable.len == 0)