Originally, the pacing code would calculate a desired `value_count_per_beat` once, at the start of the bar. This would then be the goal number of source values to process, for each beat.

However, since we pace off _source_ values, we'd often process quite a few more values than desired. This is normal and expected - we have to round up to full target value blocks, filling from the immutable table can consume a lot of values with the removal of duplicates.

Now, recalculate value_count_per_beat each beat. This self adjusts for those effects and helps to smooth out latencies better over the entire bar.

Additionally, add in a few asserts and make sure we don't exceed `beats_max` (our beat budget).

## Before
1241 batches in 52.53 s
load offered = 1000000 tx/s
load accepted = 190363 tx/s
batch latency p1 = 1 ms
batch latency p10 = 4 ms
batch latency p20 = 4 ms
batch latency p30 = 6 ms
batch latency p40 = 18 ms
batch latency p50 = 28 ms
batch latency p60 = 37 ms
batch latency p70 = 47 ms
batch latency p80 = 58 ms
batch latency p90 = 68 ms
batch latency p95 = 79 ms
batch latency p99 = 429 ms
batch latency p100 = 458 ms

## After
1247 batches in 50.53 s
load offered = 1000000 tx/s
load accepted = 197887 tx/s
batch latency p1 = 1 ms
batch latency p10 = 4 ms
batch latency p20 = 4 ms
batch latency p30 = 14 ms
batch latency p40 = 20 ms
batch latency p50 = 27 ms
batch latency p60 = 38 ms
batch latency p70 = 47 ms
batch latency p80 = 55 ms
batch latency p90 = 64 ms
batch latency p95 = 75 ms
batch latency p99 = 361 ms
batch latency p100 = 381 ms

(the effect grows even larger the more compactions there are going on!)