@cb22 2024-03-28
(cc @sentientwaffle - in #1806 you removed `beats_max`, but it's used now)